remove strict validation and add tests
